First spike of privateer.
Right now the map between id and uri is stubbed in.

But once the mapping is resolved, PATH_INFO and QUERY_STRING are
reset and reparsed, and the inner core tiddlyweb app is rerun.

This is rather tricky, but gets the job done without a bunch
conditionals and various other mappings and URI checkings. It
also means that _any_ URI the system uses can be mapped.

Of course this leaves out the hardest part: managing the mappings.

Still thinking on that one.