Add support for posting a uri to the privateer.
Send a form or JSON package with a uri field as a POST to /_
and create a tiddler in the PRIVATEER bag bag with a uuid title
and fields of uri set to the given uri and user set to the current
user. A Location with the /_/<uuid> is returned to the user.

Later when we go to to /_/<uuid> the uuid tiddler is looked for
and if found, the request is processed as if it was to the given
uri.

next:
* deleting mappings
* listing mappings