Add support for deleting mappings.
The current user must match the user named in the tiddler
for the delete to work. If there is a current user but it is
not the user named in the tiddler, then we send 404 _not_ 403
because otherwise people could farm for so-call unguessable uris
and make a catalog.

We don't want that.

next: listing