Add support for listing a user's mappings.
The output is a JSON mapping of

  full mapping uri (at the current host): full mapped uri

I experiments with using just the identifier on the left hand side,
and decided that was less useful (as you would need to be able to
reconstruct the private mapping URI from...where?)

The list is created via a filter: the tiddlers which have a
user field matching the current user.

The code now has all the basic functionality spiked in but is a bit
crufty and undocumented so next step is some refactoring.